% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvRPkgDevel-package.R
\docType{package}
\name{AdvRPkgDevel-package}
\alias{AdvRPkgDevel}
\alias{AdvRPkgDevel-package}
\title{AdvRPkgDevel: Advanced R Package Development Test Package}
\description{
So far, nothing. This package is for me to work through the process of package creation via devtools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MMMcMahon/AdvRPkgDevel}
  \item Report bugs at \url{https://github.com/MMMcMahon/AdvRPkgDevel/issues}
}

}
\author{
\strong{Maintainer}: Mike McMahon \email{Mike.McMahon@dfo-mpo.gc.ca}

}
\keyword{internal}
